// src/app/shared/types/api-types/bido-mission-equipment-dto.interface.ts
import { BidoFlightAttributeDTO } from './bido-flight-attribute-dto.interface';

export interface BidoMissionEquipmentDTO {
  missionCode?: string;
  equipmentCode?: string;
  occurrence?: number;
  customerCode?: string;
  customerOperatorCode?: string;
  missionDuration?: string;
  cycleNumber?: string;
  startDate?: number;
  endDate?: number;
  engineStartDate?: number;
  engineStopDate?: number;
  fuelUsed?: number;
  passengerNumber?: number;
  shipmentNumber?: number;
  commander?: string;
  copilot?: string;
  crewExtra?: string;
  missionComment?: string;
  missionInstanceCode?: string;
  dayMission?: boolean;
  nightMission?: boolean;
  ifrDuration?: string;
  engine1AssetCode?: string;
  engine2AssetCode?: string;
  engine3AssetCode?: string;
  engine4AssetCode?: string;
  engine1CycleNumber?: string;
  engine2CycleNumber?: string;
  engine3CycleNumber?: string;
  engine4CycleNumber?: string;
  engine1HourNumber?: string;
  engine2HourNumber?: string;
  engine3HourNumber?: string;
  engine4HourNumber?: string;
  missionArea?: string;
  plannedPassengerNumber?: number;
  deplannedPassengerNumber?: number;
  userFlightNumber?: string;
  flightLabel?: string;
  departureLocation?: string;
  arrivalLocation?: string;
  flightType?: string;
  flightDetails?: string;
  requestedOpeConfigCode?: string;
  requestedOpeConfigDescription?: string;
  requestedOpeConfigDetails?: string;
  requestedOpeConfigModifs?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  postFlightVisitDate?: number;
  postFlightVisitUser?: string;
  postFlightVisitStatus?: string;
  beforeFlightRepleDate?: string;
  beforeFlightRepleUser?: string;
  repleValidationDate?: string;
  repleValidationUser?: string;
  bidoFlightAttributeDTOList?: BidoFlightAttributeDTO[];
}