// src/app/shared/types/api-types/bido-replenishment-measure-dto.interface.ts
export interface BidoReplenishmentMeasureDTO {
  equipmentCode?: string;
  replenishmentId?: string;
  rmId?: string;
  rmNumber?: number;
  rmSource?: string;
  rmWeight?: string;
  rmVolume?: string;
  rmDensity?: string;
  rmOccurrence?: string;
  rmDate?: Date;
  rmOrigin?: string;
  engineOilReplenishmentType?: string;
  rmBatchNumber?: string;
}